requirements.txt
transformers>=5.0.0
# Install PyTorch with CUDA 12.4 support (RTX 4060 / driver 595+):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
torch>=2.6.0
accelerate>=1.1.0
datasets>=2.18.0
scikit-learn>=1.4.0
pandas>=2.2.0
gradio>=4.26.0
kagglehub>=0.3.0
pyyaml>=6.0.1
tqdm>=4.66.0
pytest>=8.1.0
beautifulsoup4>=4.12.0
pyarrow>=15.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0
nltk>=3.8.0
